Allowing DictParser to parse from a file
